fix rounding on sleepfor
Former-commit-id: d45f6185cc21787c2ae9245a614ad52ebd55e7ba

+++ src/test/test_syscoin_services.cpp
@@ -432,7 +432,7 @@ void GenerateSpendableCoins() {
 void SleepFor(const int& milliseconds, bool actualSleep) {
 	if (actualSleep)
 		MilliSleep(milliseconds);
-	int seconds = milliseconds / 1000;
+	float seconds = milliseconds / 1000;
 	BOOST_CHECK(seconds > 0);
 	UniValue r;
 	try
